Configuration (4/30)
Added configuration files and options (app/config) for both development
and production.

Added mode.php to easily and dynamically be able to switch between the
two.

Added namespace Noodlehaus\Config to start.php so that configurations can
be dynamically loaded, parsed and fetched.

Needed to add the rtrim when setting 'mode' because the Atom editor was
automatically adding a newline on save. This was causing a File Not Found
Error because Config::Load was trying to load 'development .php' NOTE the
space.