Adding DataBase Support (6/30)
Adding Database support required adding a new file database.php that is
then required by start.php.

database.php loads the Illuminate\Database\Capsule\Manager and aliases it
to just 'Capsule'. A new instance of Capsule is created and it's
properties are loaded via $app->config->get('xxx').

Capsule's bootEloquent method is called which should give us a connection
to the database and an interface to the user model.