Routes & Views (8/30)
Added Twig to be used for the Slim Views
Added Views, Routes, and Templates folders
Added a sub dir to templates to hold partial templates.
Created a named route for the home page in app/routes/home.php
Created a default template that will take the content in the view file.
The TwigExtensions provide URL helpers like urlFor and if
The view files extend the templates/default.php and define title & content