Flash Messages (9/30)
Added a route to routes/home.php for /flash.
This illustrated how to set temporary "flash" messages within the
slim framework.
The the flash 'global' message is set and then the app redirects to the
URL for home.
The flash message will appear on the home page after the redirect and then
be 'unset'