Registration (11/30)
Created dir /app/routes/auth
Created dir /app/views/auth
Created route and view files for 'register'
Created a named route for 'register.post' which handles when the request
method is a POST.
Added the param database.name to the dev and production files because it
has been missing.
The code the actually handles the insertion of a user record into the
users table is in routes/auth/register. Fields are collected using Slim's
request property and then the create method from Eloquent is user on the
user object.

After a record is inserted, a flash message is set and then the user is
redirected back to home.