Logging In (14/30)
Created functionality so that users can log in via their username or email
and password.
Added middleware that enables authentication before each request
Set Session variables to the contents of the user record on successful
logins
Appends the user data to the view object so that user methods and data can
be exposed via the views.