Sending Email 15/30
Added capability to send an email when a user successfully registers.
Sending HTML email from a view and passing data via appendData so it can
be used in the template.
The mailer is set via configuration and in this instance it's PHPMAiler
The new Mailer class creates an instance for a particular Mailer and view.
The new Message class handles methods to, subject and body.